# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "asimaster",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/push/subscribe": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Subscribe Push",
        "operationId": "subscribe_push_api_v1_push_subscribe_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Unsubscribe Push",
        "operationId": "unsubscribe_push_api_v1_push_subscribe_delete",
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushSubscriptionCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "p256dh": {
            "type": "string",
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "title": "Auth"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "endpoint",
          "p256dh",
          "auth"
        ],
        "title": "PushSubscriptionCreate"
      },
      "PushSubscriptionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "endpoint"
        ],
        "title": "PushSubscriptionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}